+++ rust/fsl-core/src/lib.rs
@@ -661,8 +661,9 @@ impl PredicateExpander {
                 hi: Box::new(self.expand_expr(*hi, &mut Vec::new())?),
                 symmetric,
             },
-            SpecItem::Struct { name, fields } => SpecItem::Struct {
+            SpecItem::Struct { name, fields, span } => SpecItem::Struct {
                 name,
+                span,
                 fields: fields
                     .into_iter()
                     .map(|(name, ty)| Ok((name, self.expand_type(ty)?)))